2006-02-10  Senganal T  <tsenganal@novell.com> 	* Mono.Data.Tds/Mono.Data.Tds.Protocol/Tds.cs : 		- GetSequentialColumn,BeginLoad, EndLoad, 		LoadData, SkipRow, SkipToColumnIndex 		New Methods to Support sequential loading of row data 	* Mono.Data.Tds/Mono.Data.Tds.Protocol/TdsComm.cs : 		- Skip : Changed arugment type from int to long. 	* System.Data/Test/ProviderTests/System.Data.SqlClient/SqlDataReaderTest.cs : 		Added more tests to verify Sequential reading of row data. 	* System.Data/Test/ProviderTests/ProviderIndependant/DataReaderTest : Corrected a failing testcase. 	* System.Data/System.Data.SqlClient/SqlDataReader.cs : 		- GetBytes : Read binary/blob/clob data sequentially when 		CommandBehavior is set to SequentialAcccess 		- GetChars : Read String/clob data sequentially when CommandBehavior 		is set to SequentialAccess 	* System.Data/System.Data.SqlClient/SqlCommand.cs : 		- ExecuteReader : set SequentialAccess property on TDS 		- CloseDataReader : Reset the command behavior
svn path=/trunk/mcs/; revision=56757

+++ mcs/class/Mono.Data.Tds/Mono.Data.Tds.Protocol/TdsComm.cs
@@ -511,7 +511,7 @@ namespace Mono.Data.Tds.Protocol {
 			}
 		}
 		
-		public void Skip (int i)
+		public void Skip (long i)
 		{
 			for ( ; i > 0; i--)
 				GetByte ();
